Extract cart utils and use it in cart related specs
