+++ plugins/woocommerce/tests/e2e-pw/tests/cart/cart.spec.js
@@ -9,115 +9,15 @@ import { addAProductToCart } from '@woocommerce/e2e-utils-playwright';
 import { expect, tags, test as baseTest } from '../../fixtures/fixtures';
 import { getFakeProduct } from '../../utils/data';
 import { createClassicCartPage, CLASSIC_CART_PAGE } from '../../utils/pages';
+import { checkCartContent } from '../../utils/cart';
 import { updateIfNeeded, resetValue } from '../../utils/settings';
 import { WC_API_PATH } from '../../utils/api-client';
 
 const cartPages = [ { name: 'blocks cart', slug: 'cart' }, CLASSIC_CART_PAGE ];
 
-/* region helpers */
 function isClassicCart( page ) {
 	return page.url().includes( CLASSIC_CART_PAGE.slug );
 }
-
-function formatAmount( amount ) {
-	return parseFloat( amount ).toLocaleString( 'en-US', {
-		minimumFractionDigits: 2,
-		maximumFractionDigits: 2,
-	} );
-}
-
-async function checkCartContentInClassicCart(
-	page,
-	products,
-	tax,
-	expectedTotal
-) {
-	for ( const product of products ) {
-		const row = await page
-			.locator( 'tr.cart_item' )
-			.filter( { hasText: product.data.name } );
-
-		await expect( row ).toHaveCount( 1 );
-		await expect( row.getByLabel( 'Product quantity' ) ).toHaveValue(
-			product.qty.toString()
-		);
-
-		await expect( row.locator( 'td.product-price' ) ).toHaveText(
-			`$${ formatAmount( product.data.price ) }`
-		);
-	}
-
-	const totalRow = page.locator( '.cart_totals tr.order-total' );
-
-	await expect( totalRow.locator( 'td' ) ).toHaveText(
-		`$${ formatAmount( expectedTotal ) }`
-	);
-}
-
-async function checkCartContentInBlocksCart(
-	page,
-	products,
-	tax,
-	expectedTotal
-) {
-	for ( const product of products ) {
-		const row = await page
-			.locator( 'tr.wc-block-cart-items__row' )
-			.filter( { hasText: product.data.name } );
-
-		await expect( row ).toHaveCount( 1 );
-		await expect(
-			row.getByRole( 'spinbutton', { name: 'Quantity' } )
-		).toHaveValue( product.qty.toString() );
-
-		const expectedItemTotal = formatAmount(
-			parseFloat( product.data.price ) * product.qty
-		);
-		await expect(
-			row.locator( 'td.wc-block-cart-item__total' )
-		).toHaveText( `$${ expectedItemTotal }` );
-	}
-
-	await expect(
-		page.locator( '.wc-block-components-totals-item__value' ).last()
-	).toHaveText( `$${ formatAmount( expectedTotal ) }` );
-}
-
-async function checkCartContent( slug, page, products, tax ) {
-	if ( products.length === 0 ) {
-		await expect(
-			page.locator( 'main' ).getByText( 'Your cart is currently empty' )
-		).toBeVisible();
-		return;
-	}
-
-	const expectedTotal = products.reduce( ( total, product ) => {
-		return (
-			total +
-			parseFloat( product.data.price ) *
-				product.qty *
-				( 1 + parseFloat( tax.rate ) / 100 )
-		);
-	}, 0 );
-
-	if ( isClassicCart( page ) ) {
-		await checkCartContentInClassicCart(
-			page,
-			products,
-			tax,
-			expectedTotal
-		);
-	} else {
-		await checkCartContentInBlocksCart(
-			page,
-			products,
-			tax,
-			expectedTotal
-		);
-	}
-}
-
-/* endregion */
 
 /* region fixtures */
 const test = baseTest.extend( {
@@ -229,14 +129,14 @@ cartPages.forEach( ( { name, slug } ) => {
 		async ( { page, products, tax } ) => {
 			await test.step( 'empty cart is displayed', async () => {
 				page.goto( slug );
-				await checkCartContent( slug, page, [], tax );
+				await checkCartContent( isClassicCart( page ), page, [], tax );
 			} );
 
 			await test.step( 'one product in cart is displayed', async () => {
 				await addAProductToCart( page, products[ 0 ].id, 2 );
 				await page.goto( slug );
 				await checkCartContent(
-					slug,
+					isClassicCart( page ),
 					page,
 					[ { data: products[ 0 ], qty: 2 } ],
 					tax
@@ -254,7 +154,7 @@ cartPages.forEach( ( { name, slug } ) => {
 						.fill( '3' );
 				}
 				await checkCartContent(
-					slug,
+					isClassicCart( page ),
 					page,
 					[ { data: products[ 0 ], qty: 3 } ],
 					tax
@@ -265,7 +165,7 @@ cartPages.forEach( ( { name, slug } ) => {
 				await addAProductToCart( page, products[ 1 ].id, 1 );
 				await page.goto( slug );
 				await checkCartContent(
-					slug,
+					isClassicCart( page ),
 					page,
 					[
 						{ data: products[ 0 ], qty: 3 },
@@ -284,7 +184,7 @@ cartPages.forEach( ( { name, slug } ) => {
 				).toBeVisible();
 				await page.goBack();
 				await checkCartContent(
-					slug,
+					isClassicCart( page ),
 					page,
 					[
 						{ data: products[ 0 ], qty: 3 },
@@ -310,7 +210,7 @@ cartPages.forEach( ( { name, slug } ) => {
 					.click();
 
 				await checkCartContent(
-					slug,
+					isClassicCart( page ),
 					page,
 					[ { data: products[ 1 ], qty: 1 } ],
 					tax
@@ -331,7 +231,7 @@ cartPages.forEach( ( { name, slug } ) => {
 					)
 					.click();
 
-				await checkCartContent( slug, page, [], tax );
+				await checkCartContent( isClassicCart( page ), page, [], tax );
 			} );
 		}
 	);
